ecosystem/intelligence/Praxoimous-AI-Gateway-MVP/api/v1/endpoints/rag_interface_router.py
# api/v1/endpoints/rag_interface_router.py
from fastapi import APIRouter, Depends, HTTPException, status
from pydantic import BaseModel
from typing import Dict, Any

from core.logger import log
from core.license_manager import is_feature_enabled, Feature, get_current_license_tier

router = APIRouter()

async def verify_rag_access():
    """Dependency to check if the RAG feature is enabled by the license."""
    if not is_feature_enabled(Feature.RAG_INTERFACE):
        current_tier_name = get_current_license_tier().name
        log.warning(f"RAG Interface access denied. Current tier: {current_tier_name}. Required for feature: {Feature.RAG_INTERFACE.name}")
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail=f"The RAG Interface feature is not available for your current license tier ({current_tier_name}). Please upgrade to Enterprise."
        )

class RAGQueryRequest(BaseModel):
    query: str
    collection: str = "default"

@router.post("/rag/query", summary="Query RAG Interface", dependencies=[Depends(verify_rag_access)])
async def query_rag_interface(request: RAGQueryRequest):
    """
    (Placeholder) Endpoint for querying the Retrieval-Augmented Generation interface.
    This is an Enterprise-level feature.
    """
    log.info(f"RAG query received for collection '{request.collection}': '{request.query[:50]}...'")
    # In a real implementation, this would query a vector database and then pass the results to an LLM.
    return {"status": "success", "message": "RAG query processed (placeholder).", "result": f"This is a placeholder response for the query: '{request.query}'"}

@router.get("/rag/settings", summary="Get RAG Settings", dependencies=[Depends(verify_rag_access)])
async def get_rag_settings() -> Dict[str, Any]:
    """
    (Placeholder) Endpoint for retrieving RAG configuration and settings.
    """
    return {"status": "success", "settings": {"vector_db": "chroma", "embedding_model": "text-embedding-ada-002", "collections": ["default", "financials"]}}