ecosystem/intelligence/Praxoimous-AI-Gateway-MVP/requirements.txt
# requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.23.2
PyYAML==6.0.1
pytest==7.4.2
httpx==0.25.0
pytest-mock==3.11.1
google-generativeai==0.5.4 # Or the latest version
beautifulsoup4==4.12.3 # Or the latest version
pytz==2023.3.post1 # Or the latest version
vaderSentiment==3.3.2 # For SentimentAnalysisSkill
cryptography==42.0.5 # For license key generation and verification